bug: Erreur sur la suppression d'un fichier
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@15795 a21d6be2-e512-0410-a060-91f5a2946631

+++ modules/system/ajax_dispatch_files.php
@@ -50,7 +50,9 @@ $count = CAppUI::conf("eai max_files_to_process");
 $files = array_slice($files, 0, $count);
 
 // Mise à jour du fichier avec le nouveau diff
-unlink($filename_excludes);
+if (file_exists($filename_excludes)) {
+  unlink($filename_excludes);
+}  
 $file  = fopen($filename_excludes, "a+");
 foreach (array_diff($files_excludes, $array_diff) as $_file_exclude) {
   fputs($file, $_file_exclude."\n");
